requirements.txt
Django (>=1.7)
six
infi.clickhouse-orm
huey
